mudança devolver item e forma de implementar multa

+++ PJBL9-Biblioteca/Biblioteca/Biblioteca/Livro.java
@@ -14,7 +14,7 @@ public class Livro extends itemBiblioteca{
     }
 
     @Override
-    public float calculaMulta(int diasAtraso) {
+    public float calculaMulta(long diasAtraso) {
         float taxaDiaria = (float) 1;
         float multaLivro = diasAtraso * taxaDiaria;
         return multaLivro;
